requirements.txt
Django==5.2.*
django-tailwind-cli
djangorestframework
django-debug-toolbar
django-unfold
django-cotton
django-pwa

pillow
python-decouple

gunicorn
psycopg2-binary
